refactor: consolidate KeygenPendingState and enhance KeygenFlow with new components
- Replaced the existing `KeygenPendingState` component with a new implementation that integrates `KeygenProductEducation` and `KeygenProgressIndicator`, improving the user experience during the key generation process.
- Updated `JoinKeygenProcess` and `KeygenFlow` to utilize the new `KeygenPendingState`, streamlining the pending state handling and enhancing maintainability.
- Removed obsolete `KeygenPendingState` and `SlidesLoader` components to clean up the codebase and reduce complexity.
