+++ clients/desktop/src/vault/keygen/join/JoinKeygenProcess.tsx
@@ -1,4 +1,5 @@
 import { useKeygenMutation } from '@core/ui/mpc/keygen/mutations/useKeygenMutation'
+import { KeygenPendingState } from '@core/ui/mpc/keygen/progress/KeygenPendingState'
 import { CurrentVaultProvider } from '@core/ui/vault/state/currentVault'
 import { Match } from '@lib/ui/base/Match'
 import { StepTransition } from '@lib/ui/base/StepTransition'
@@ -9,11 +10,9 @@ import { useEffect } from 'react'
 import { useAppNavigate } from '../../../navigation/hooks/useAppNavigate'
 import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { BackupSecureVault } from '../../setup/secure/backup/BackupSecureVault'
-import { SetupVaultEducationSlides } from '../../setup/shared/SetupVaultCreationStep/SetupVaultEducationSlides'
 import { SetupVaultSuccessScreen } from '../../setup/shared/SetupVaultSuccessScreen'
 import { KeygenFailedState } from '../shared/KeygenFailedState'
 import { KeygenPageHeader } from '../shared/KeygenPageHeader'
-import { KeygenPendingState } from '../shared/KeygenPendingState'
 import { KeygenSuccessStep } from '../shared/KeygenSuccessStep'
 
 export const JoinKeygenProcess = ({
@@ -65,17 +64,10 @@ export const JoinKeygenProcess = ({
         </>
       )}
       pending={() => (
-        <Match
-          value={keygenType}
-          create={() => <SetupVaultEducationSlides value={step} />}
-          migrate={() => <SetupVaultEducationSlides value={step} />}
-          reshare={() => (
-            <>
-              <KeygenPageHeader title={title} />
-              <KeygenPendingState value={step} />
-            </>
-          )}
-        />
+        <>
+          <KeygenPageHeader title={title} />
+          <KeygenPendingState value={step} />
+        </>
       )}
     />
   )
