+++ core/ui/mpc/keygen/progress/KeygenPendingState.tsx
@@ -1,15 +1,13 @@
 import { KeygenStep } from '@core/mpc/keygen/KeygenStep'
-import { CreateVaultDynamicEducationContent } from '@core/ui/mpc/keygen/create/education/CreateVaultDynamicEducationalContent'
+import { KeygenProgressIndicator } from '@core/ui/mpc/keygen/progress/KeygenProgressIndicator'
 import { VStack } from '@lib/ui/layout/Stack'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { ValueProp } from '@lib/ui/props'
 import styled from 'styled-components'
 
-import { SlidesLoader } from './components/SlidesLoader'
+import { KeygenProductEducation } from '../education/product/KeygenProductEducation'
 
-export const SetupVaultEducationSlides = ({
-  value,
-}: ValueProp<KeygenStep | null>) => {
+export const KeygenPendingState = ({ value }: ValueProp<KeygenStep | null>) => {
   return (
     <PageContent>
       <Wrapper
@@ -19,9 +17,9 @@ export const SetupVaultEducationSlides = ({
         gap={48}
       >
         <VStack flexGrow justifyContent="center">
-          <CreateVaultDynamicEducationContent />
+          <KeygenProductEducation />
         </VStack>
-        <SlidesLoader value={value} />
+        <KeygenProgressIndicator value={value} />
       </Wrapper>
     </PageContent>
   )
